requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
google-generativeai>=0.5.0
python-dotenv==1.0.1
Pillow==10.0.0
requests==2.31.0
beautifulsoup4==4.12.0
validators==0.22.0
pytesseract==0.3.10
python-magic-bin==0.4.14
aiohttp==3.9.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Flask-WTF==1.2.1
Flask-Limiter==3.5.1
Flask-Caching==2.1.0
Flask-Talisman==1.1.0
structlog==24.1.0
# ASGI server and dependencies
uvicorn==0.25.0
asgiref==3.7.2
h11==0.14.0
httptools==0.6.1
a2wsgi==1.10.0  # WSGI to ASGI adapter for Flask
# Optional: Alternative ASGI server
# hypercorn==0.16.0